优化项目代码结构，解决在表格下面首行第一个字符前删除时会与表格合并的问题

+++ src/customEditor/withTables.js
@@ -1,39 +1,49 @@
-import { Range, Editor, Point, Transforms } from "slate"
+import { Range, Editor, Point, Transforms, Node } from "slate"
 
 const withTables = (editor) => {
     const { deleteBackward, deleteForward, insertBreak } = editor
 
-    editor.deleteBackward = (unit) => {
+    const commonDelete = () => {
         const { selection } = editor
         if (selection && Range.isCollapsed(selection)) {
 
-            // 需要实现的功能：在表格下面删除空元素时不进行合并
-            const element = Editor.node(editor, selection);
-            const isEmpty = Editor.isEmpty(editor, element);
-            console.log(isEmpty)
-            const prevNode = Editor.previous(editor);
-            if (prevNode) {
-                const [, prevNodePath] = prevNode;
-                const parent = Editor.parent(editor, prevNodePath);
-                if (parent) {
-                    const [{ type }] = parent;
-                    if (type === 'tableAfter') {
+            // ### 实现的功能：在表格下面删除空元素时不进行合并 ###
 
-                        return
+            // 判断光标是否在位置0
+            if (selection.anchor.offset === 0) {
+                // 查找上一个节点
+                const prevNode = Editor.previous(editor, { at: selection });
+                if (prevNode) {
+                    const [prevTableAfter] = Editor.nodes(editor, {
+                        at: prevNode[1],
+                        match: n => n.type === 'tableAfter',
+                    });
+                    // 如果上一个节点是table, 则阻止默认删除
+                    if (prevTableAfter) {
+                        const node = editor.children[selection.anchor.path[0]];
+                        // 如果节点为空，则删除节点
+                        if (Node.string(node).length === 0) {
+                            Transforms.removeNodes(editor, {
+                                at: selection
+                            })
+                        }
+                        return true;
                     }
                 }
             }
 
+            // 判断是否是表格元素
             const [tableWrap] = Editor.nodes(editor, {
                 match: (n) => n.type === 'tableWrap'
             });
 
             if (tableWrap) {
+                // ### 实现的功能： 如果是在表格之前删除，则阻止删除，如果是在表格之后进行删除，则删除表格
                 const [tableBefore] = Editor.nodes(editor, {
                     match: (n) => n.type === 'tableBefore'
                 })
                 if (tableBefore) {
-                    return
+                    return true;
                 }
 
                 const [tableAfter] = Editor.nodes(editor, {
@@ -43,7 +53,7 @@ const withTables = (editor) => {
                     Transforms.removeNodes(editor, {
                         match: n => n.type === 'tableWrap'
                     })
-                    return
+                    return true;
                 }
     
     
@@ -56,40 +66,30 @@ const withTables = (editor) => {
                     const start = Editor.start(editor, cellPath)
     
                     if (Point.equals(selection.anchor, start)) {
-                        return
+                        return true;
                     }
                 }
             }
         }
+        return false;
+    }
 
+    editor.deleteBackward = (unit) => {
+        if (commonDelete()) return
         deleteBackward(unit)
     }
 
     editor.deleteForward = (unit) => {
-        const { selection } = editor
-        console.log(unit)
-        if (selection && Range.isCollapsed(selection)) {
-            const [cell] = Editor.nodes(editor, {
-                match: (n) => n.type === 'td'
-            })
-
-            if (cell) {
-                const [, cellPath] = cell
-                const end = Editor.end(editor, cellPath)
-
-                if (Point.equals(selection.anchor, end)) {
-                    return
-                }
-            }
-        }
-
+        if (commonDelete()) return
         deleteForward(unit)
     }
 
     editor.insertBreak = () => {
         const { selection } = editor
+        const text = { text: '' };
+        const p = { type: 'p', children: [text] };
 
-        if (selection) {
+        if (selection && Range.isCollapsed(selection)) {
             const [tableBefore] = Editor.nodes(editor, {
                 match: (n) => n.type === 'tableBefore'
             })
@@ -101,8 +101,6 @@ const withTables = (editor) => {
                 match: (n) => n.type === 'tableAfter'
             })
             if (tableAfter) {
-                const text = { text: '' };
-                const p = { type: 'p', children: [text] };
                 Transforms.insertNodes(editor, p, {
                     match: n => n.type === 'tableWrap'
                 });
@@ -112,6 +110,10 @@ const withTables = (editor) => {
             const [table] = Editor.nodes(editor, { match: (n) => n.type === 'table' })
 
             if (table) {
+                const [td] = Editor.nodes(editor, { match: (n) => n.type === 'td' })
+                if (td) {
+                    Transforms.insertNodes(editor, p);
+                }
                 return
             }
         }
